removed/renamed useless TestColorSense/ColorSensor
