+++ 2020FRCsoftware/src/main/java/frc/robot/subsystems/ColorSensor.java
@@ -1,5 +1,7 @@
+//Not entirely sure how this will work, it's an example.
+
 /*----------------------------------------------------------------------------*/
-/* Copyright (c) 2018 FIRST. All Rights Reserved.                             */
+/* Copyright (c) 2017-2018 FIRST. All Rights Reserved.                        */
 /* Open Source Software - may be modified and shared by FRC teams. The code   */
 /* must be accompanied by the FIRST BSD license file in the root directory of */
 /* the project.                                                               */
@@ -7,26 +9,71 @@
 
 package frc.robot.subsystems;
 
-import edu.wpi.first.wpilibj.command.Subsystem;
-import frc.robot.commands.ColorCommand;
+import edu.wpi.first.wpilibj.TimedRobot;
+import edu.wpi.first.wpilibj.I2C;
+import edu.wpi.first.wpilibj.smartdashboard.SmartDashboard;
+import edu.wpi.first.wpilibj.util.Color;
 
 import com.revrobotics.ColorSensorV3;
 
+public class ColorSensor extends TimedRobot {
+  /**
+   * Change the I2C port below to match the connection of your color sensor
+   */
+  private final I2C.Port i2cPort = I2C.Port.kOnboard;
 
-/**
- * Add your docs here.
- */
-
-public class ColorSensor extends Subsystem {
-
-  // Put methods for controlling this subsystem
-  // here. Call these from Commands.
+  /**
+   * A Rev Color Sensor V3 object is constructed with an I2C port as a 
+   * parameter. The device will be automatically initialized with default 
+   * parameters.
+   */
+  private final ColorSensorV3 m_colorSensor = new ColorSensorV3(i2cPort);
 
   @Override
-  public void initDefaultCommand() {
-    setDefaultCommand(new ColorCommand());
+  public void robotPeriodic() {
+    /**
+     * The method GetColor() returns a normalized color value from the sensor and can be
+     * useful if outputting the color to an RGB LED or similar. To
+     * read the raw color, use GetRawColor().
+     * 
+     * The color sensor works best when within a few inches from an object in
+     * well lit conditions (the built in LED is a big help here!). The farther
+     * an object is the more light from the surroundings will bleed into the 
+     * measurements and make it difficult to accurately determine its color.
+     */
+    Color detectedColor = m_colorSensor.getColor();
 
-    // Set the default command for a subsystem here.
-    // setDefaultCommand(new initColorSensor());
+    /**
+     * The sensor returns a raw IR value of the infrared light detected.
+     */
+    double IR = m_colorSensor.getIR();
+
+    /**
+     * Open Smart Dashboard or Shuffleboard to see the color detected by the 
+     * sensor.
+     */
+    SmartDashboard.putNumber("Red", detectedColor.red);
+    SmartDashboard.putNumber("Green", detectedColor.green);
+    SmartDashboard.putNumber("Blue", detectedColor.blue);
+    SmartDashboard.putNumber("IR", IR);
+
+    /**
+     * In addition to RGB IR values, the color sensor can also return an 
+     * infrared proximity value. The chip contains an IR led which will emit
+     * IR pulses and measure the intensity of the return. When an object is 
+     * close the value of the proximity will be large (max 2047 with default
+     * settings) and will approach zero when the object is far away.
+     * 
+     * Proximity can be used to roughly approximate the distance of an object
+     * or provide a threshold for when an object is close enough to provide
+     * accurate color values.
+     */
+
+     /**   
+      * int proximity = m_colorSensor.getProximity();
+        SmartDashboard.putNumber("Proximity", proximity);
+    */
+  
   }
 }
+
